Changes for 5.7.7beta03

+++ src/mbio/mbsys_singlebeam.c
@@ -226,7 +226,7 @@ int mbsys_singlebeam_pingnumber(int verbose, void *mbio_ptr, unsigned int *pingn
 	return (status);
 }
 /*--------------------------------------------------------------------*/
-int mbsys_singlebeam_segynumber(int verbose, void *mbio_ptr, int *line, int *shot, int *cdp, int *error) {
+int mbsys_singlebeam_segynumber(int verbose, void *mbio_ptr, unsigned int *line, unsigned int *shot, unsigned int *cdp, int *error) {
 	if (verbose >= 2) {
 		fprintf(stderr, "\ndbg2  MBIO function <%s> called\n", __func__);
 		fprintf(stderr, "dbg2  Input arguments:\n");
@@ -241,9 +241,9 @@ int mbsys_singlebeam_segynumber(int verbose, void *mbio_ptr, int *line, int *sho
 	struct mbsys_singlebeam_struct *store = (struct mbsys_singlebeam_struct *)mb_io_ptr->store_data;
 
 	/* extract data from structure */
-	*line = store->seismic_line;
-	*shot = store->seismic_shot;
-	*cdp = store->seismic_cdp;
+	*line = (unsigned int) store->seismic_line;
+	*shot = (unsigned int) store->seismic_shot;
+	*cdp = (unsigned int) store->seismic_cdp;
 
 	const int status = MB_SUCCESS;
